Rename eventtarget.js to event-target.js
